@@ -1,22 +1,18 @@
 package model
 
 import (
-	"encoding"
+	"errors"
 	"fmt"
 	"io"
 	"strconv"
 	"time"
-
-	"github.com/99designs/gqlgen/graphql"
 )
 
 type ISO8601DateTime time.Time
 
 var (
-	_ fmt.Stringer             = new(ISO8601DateTime)
-	_ encoding.TextUnmarshaler = new(ISO8601DateTime)
-	_ graphql.Marshaler        = new(ISO8601DateTime)
-	_ graphql.Unmarshaler      = new(ISO8601DateTime)
+	ErrParsingISO8601DateTime           = errors.New("error parsing ISO8601DateTime")
+	ErrUnableToUnmarshalISO8601FromType = errors.New("error parsing ISO8601DateTime: value must be string")
 )
 
 // String implements fmt.Stringer.
@@ -28,7 +24,7 @@ func (dt ISO8601DateTime) String() string {
 func (dt *ISO8601DateTime) UnmarshalText(text []byte) error {
 	parsed, err := time.Parse(time.RFC3339Nano, string(text))
 	if err != nil {
-		return fmt.Errorf("error parsing ISO8601DateTime: %v", err)
+		return fmt.Errorf("%w: %v", ErrParsingISO8601DateTime, err)
 	}
 	*dt = ISO8601DateTime(parsed)
 	return nil
@@ -42,7 +38,7 @@ func (dt *ISO8601DateTime) UnmarshalGQL(v any) error {
 	if value, ok := v.([]byte); ok {
 		return dt.UnmarshalText(value)
 	}
-	return fmt.Errorf("error parsing ISO8601DateTime: unexpected value type %T", v)
+	return fmt.Errorf("%w, got %T", ErrUnableToUnmarshalISO8601FromType, v)
 }
 
 // MarshalGQL implements graphql.Marshaler.
